+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project are documented in this file.
 The format adheres to Semantic Versioning.
 
+## 0.3.3 (2018-01-16): Basic Databook Items
+
+Features
+- Improved wrapped `createDatabookFunc()` to produce a list of populations in a databook.
+
 ## 0.3.2 (2018-01-16): Initial Databook Construction
 
 Features